finishing

+++ database/migrations/2023_09_04_144536_create_products_table.php
@@ -17,7 +17,7 @@ return new class extends Migration
             $table->string('name');
             $table->unsignedBigInteger('category')->nullable();
             $table->foreign('category')->references('id')->on('categories')->onDelete('cascade');
-            $table->decimal('price', 8, 2);
+            $table->bigInteger('price'); // Corrected column type
             $table->text('description');
             $table->integer('quantity');
             $table->timestamps();
